.aff Census Search for All Gulf Coast Counties
This is an .aff file that can be loaded to the "Your Selection" search window on the left of the screen when in https://factfinder.census.gov.  It will load all of the county names so they don't need to be manually entered when searching for and downloading a census varaible from https://factfinder.census.gov.